Moving 'courses' to new admin page, pt. 2
